requirements.txt
# langchain-core==1.2.16
# langchain==1.2.10
# langchain-openai==1.1.10
# streamlit==1.54.0
# beautifulsoup4==4.14.3
# fastapi==0.133.1
# html5lib==1.1
# jinja2==3.1.6
# langchain==1.2.10
# langchain-astradb==1.0.0
# langchain_core==1.2.16
# langchain-google-genai==4.2.1
# langchain-groq==1.1.2
# lxml==6.0.2
# python-dotenv==1.2.1
# python-multipart==0.0.22
# selenium==4.41.0
# streamlit==1.54.0
# undetected-chromedriver==3.5.5
# uvicorn==0.41.0
# structlog==25.5.0
# langgraph==1.0.9
# ragas==0.4.3
# langchain-mcp-adapters==0.2.1
# mcp==1.26.0
# ddgs==9.10.0
# langchain-openai==1.1.10

-e .

langchain-core==0.1.53
langchain==0.1.20
langchain-openai==0.1.7
streamlit==1.54.0
beautifulsoup4==4.14.3
fastapi==0.133.1
html5lib==1.1
jinja2==3.1.6
langchain==0.1.20
langchain-astradb==1.0.0
langchain_core==0.1.53
langchain-google-genai==1.0.3
langchain-groq==0.1.4
lxml==6.0.2
python-dotenv==1.2.1
python-multipart==0.0.22
selenium==4.41.0
streamlit==1.54.0
undetected-chromedriver==3.5.5
uvicorn==0.41.0
structlog==25.5.0
langgraph==0.0.55
ragas==0.4.3
langchain-mcp-adapters==0.2.1
mcp==1.26.0
ddgs==9.10.0
langchain-openai==0.1.7